Fit a DFA model with TMB.

Fit a DFA model with TMB.

dfaTMB(
  obs,
  NumStates = 1,
  ErrStruc = "DE",
  EstCovar = FALSE,
  Covars = NULL,
  indivCovar = FALSE,
  Dmat = NULL,
  Dfac = NULL,
  Rfac = NULL,
  logsdObs = NULL,
  logsdObsFac = NULL,
  cholCorr = NULL,
  cholFac = NULL,
  EstSE = FALSE,
  silent = TRUE
)

Arguments

obs

Vector of observations n x T.

NumStates

m

ErrStruc

DE=diagonal and equal, UNC=unconstrained, DUE=diagonal and unequal

EstCovar

TRUE/FALSE

silent

Show TMB output when fitting

Value

A list with Optimization, Estimates, Fits, and AIC

Author

Tim Cline wrote most of this while a graduate student in the Fish 507 Time Series Analysis course. Eli Holmes later modified it to replicate the MARSS(x, form="dfa") model.

Examples

library(MARSS)
data(lakeWAplankton, package = "MARSS")
phytoplankton <- c("Cryptomonas", "Diatoms", "Greens", "Unicells", "Other.algae")
dat <- as.data.frame(lakeWAplanktonTrans) |>
  subset(Year >= 1980 & Year <= 1989) |>
  subset(select=phytoplankton) |>
  t() |>
  MARSS::zscore()

#fit with MARSS
m1.em <- MARSS(dat, model=list(R='unconstrained', m=1, tinitx=1), form='dfa', z.score=FALSE)
#> Success! abstol and log-log tests passed at 449 iterations.
#> Alert: conv.test.slope.tol is 0.5.
#> Test with smaller values (<0.1) to ensure convergence.
#> 
#> MARSS fit is
#> Estimation method: kem 
#> Convergence test: conv.test.slope.tol = 0.5, abstol = 0.001
#> Estimation converged in 449 iterations. 
#> Log-likelihood: -772.4017 
#> AIC: 1584.803   AICc: 1586.264   
#>  
#>         Estimate
#> Z.11      0.1158
#> Z.21      0.1239
#> Z.31      0.0898
#> Z.41      0.3919
#> Z.51     -0.1652
#> R.(1,1)   0.9364
#> R.(2,1)   0.3157
#> R.(3,1)   0.1475
#> R.(4,1)   0.2863
#> R.(5,1)   0.2006
#> R.(2,2)   0.9284
#> R.(3,2)  -0.0404
#> R.(4,2)   0.0636
#> R.(5,2)  -0.0233
#> R.(3,3)   0.9641
#> R.(4,3)   0.0314
#> R.(5,3)   0.2919
#> R.(4,4)   0.3587
#> R.(5,4)   0.2605
#> R.(5,5)   0.8792
#> Initial states (x0) defined at t=1
#> 
#> Standard errors have not been calculated. 
#> Use MARSSparamCIs to compute CIs and bias estimates.
#> 
mym1 <- dfaTMB(obs = dat, NumStates = 1, ErrStruc='UNC')
#> initial  value 772.401093 
#> iter   2 value 772.401093
#> iter   3 value 772.401093
#> iter   3 value 772.401093
#> iter   3 value 772.401093
#> final  value 772.401093 
#> converged
cbind(coefficients(m1.em)$R,as.vector(mym1$Estimates$R[lower.tri(mym1$Estimates$R,diag=T)]))
#>              [,1]        [,2]
#> (1,1)  0.93638019  0.93666716
#> (2,1)  0.31573710  0.31613534
#> (3,1)  0.14751952  0.14774356
#> (4,1)  0.28632693  0.28713032
#> (5,1)  0.20062137  0.20043162
#> (2,2)  0.92838765  0.92890853
#> (3,2) -0.04039285 -0.04006646
#> (4,2)  0.06363202  0.06478714
#> (5,2) -0.02331011 -0.02363784
#> (3,3)  0.96406442  0.96428513
#> (4,3)  0.03142312  0.03213861
#> (5,3)  0.29186021  0.29169763
#> (4,4)  0.35870272  0.36081510
#> (5,4)  0.26046939  0.26008360
#> (5,5)  0.87924072  0.87919006
cbind(coefficients(m1.em)$Z,as.vector(mym1$Estimates$Z)) 
#>           [,1]       [,2]
#> 11  0.11581668  0.1147769
#> 21  0.12390577  0.1226094
#> 31  0.08978938  0.0889083
#> 41  0.39187823  0.3887298
#> 51 -0.16515621 -0.1641403